move files to links/model
