+++ chainercv/links/__init__.py
@@ -1,2 +1,2 @@
-from chainercv.links.ssd import SSD300  # NOQA
-from chainercv.links.ssd import SSD512  # NOQA
+from chainercv.links.model.ssd import SSD300  # NOQA
+from chainercv.links.model.ssd import SSD512  # NOQA
