+++ chainercv/links/model/ssd/__init__.py
@@ -1,0 +1,7 @@
+from chainercv.links.model.ssd.normalize import Normalize  # NOQA
+from chainercv.links.model.ssd.ssd import SSD  # NOQA
+from chainercv.links.model.ssd.ssd_vgg16 import VGG16Extractor  # NOQA
+from chainercv.links.model.ssd.ssd_vgg16 import VGG16Extractor300  # NOQA
+from chainercv.links.model.ssd.ssd_vgg16 import VGG16Extractor512  # NOQA
+from chainercv.links.model.ssd.ssd_vgg16 import SSD300  # NOQA
+from chainercv.links.model.ssd.ssd_vgg16 import SSD512  # NOQA
