+++ chainercv/links/model/ssd/normalize.py
