--- chainercv/links/ssd/__init__.py
@@ -1,4 +1,0 @@
-from chainercv.links.ssd.normalize import Normalize  # NOQA
-from chainercv.links.ssd.ssd import SSD  # NOQA
-from chainercv.links.ssd.ssd_vgg16 import SSD300  # NOQA
-from chainercv.links.ssd.ssd_vgg16 import SSD512  # NOQA
